+++ scripts/output/projects/FSDP_collect.R
@@ -119,24 +119,6 @@ for (i in 1:length(outputdir)) {
     } else missing <- c(missing,outputdir[i])
 
     ## Water
-    nc_file <- file.path(outputdir[i], paste("efvViolation.mz", sep = "-"))
-    if (file.exists(nc_file)) {
-      a <- read.magpie(nc_file)[, years, ]
-      getNames(a) <- "environmental flow violations"
-      getSets(a,  fulldim = FALSE)[3] <- "variable"
-      a <- addLocation(a)
-      y <- mbind(y, a)
-    } else missing <- c(missing, outputdir[i])
-
-    nc_file <- file.path(outputdir[i], paste("watStress.mz", sep = "-"))
-    if (file.exists(nc_file)) {
-      a <- read.magpie(nc_file)[, years, ]
-      getNames(a) <- "water withdrawal to availability ratio (ratio)"
-      getSets(a,  fulldim = FALSE)[3] <- "variable"
-      a <- addLocation(a)
-      y <- mbind(y, a)
-    } else missing <- c(missing, outputdir[i])
-
     nc_file <- file.path(outputdir[i], paste("watStressViolations.mz", sep = "-"))
     if (file.exists(nc_file)) {
       a <- read.magpie(nc_file)[, years, ]