+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### changed
-- **13_tc** relaxed vm_tau upper limit
+- **inputs** updated f56_emis_policy
 - **scripts** updated FSEC start and output scripts
 - **scripts** update of rds_report to allow gridded intermediate outputs
 - **config** non-food system emission MAGICC switch
@@ -21,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **scripts** log files are now written in a subfolder "logs"
 
 ### added
-- **scripts** added water output script for FSEC model runs 
+- **scripts** added water output script for FSEC model runs
 - **scripts** added output script running MAGICC7 on a MAgPIE scenario
 - **scripts** added output script for gridded crop diversity indices
 - **scripts** added output scripts for FSEC FSDP runs